Svg correcting

+++ PcbReader/src/Layers/Svg/Entities/PathParts/ArcSvgPathPart.cs
@@ -5,7 +5,7 @@ namespace PcbReader.Layers.Svg.Entities;
 
 public class ArcSvgPathPart : ISvgPathPart, ISvgCursorDriver {
     public Point PointTo { get; set; }
-    public decimal Radius { get; set; }
+    public double Radius { get; set; }
     public bool IsLargeArc { get; set; } = false;
     public RotationDirection RotationDirection { get; set; } = 0;
 }
